MindGame
In this memory game, there are 16 closed boxes, each containing two of every letter from 'a' to 'h'. As the player correctly guesses the pairs, the boxes open. The game ends when all pairs are correctly guessed. Additionally, with serialization, the player can save their progress and continue later or start a new game.